requirements.txt
absl-py==2.1.0
asarPy==1.0.1
beautifulsoup4==4.12.3
boto3==1.34.139
botocore==1.34.139
bs4==0.0.2
certifi==2024.6.2
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
Cython==0.29.37
decorator==4.4.2
defusedxml==0.7.1
easyocr==1.7.1
engineering-notation==0.10.0
filelock==3.15.4
fire==0.6.0
fonttools==4.53.0
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.64.1
huggingface-hub==0.23.4
idna==3.7
imageio==2.34.2
imageio-ffmpeg==0.5.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
kiwisolver==1.4.5
lazy_loader==0.4
Mako==1.3.5
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.9.1
meson==1.3.2
moviepy==1.0.3
mpmath==1.3.0
networkx==3.3
ninja==1.11.1.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
opencv-python==4.8.1.78
opencv-python-headless==4.10.0.84
packaging==24.1
pandas==2.2.2
pillow==10.4.0
proglog==0.1.10
protobuf==4.25.3
psutil==6.0.0
py-cpuinfo==9.0.0
pybboxes==0.1.6
pyclipper==1.3.0.post5
Pygments==2.18.0
pyparsing==3.1.2
pytesseract==0.3.10
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-magic==0.4.27
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
requests-toolbelt==1.0.0
roboflow==1.1.34
s3transfer==0.10.2
sahi==0.11.16
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
shapely==2.0.4
six==1.16.0
smmap==5.0.1
soupsieve==2.5
supervision==0.21.0
supervisor==4.2.5
sympy==1.12.1
tensorboard==2.17.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tifffile==2024.7.2
tk==0.1.0
tk-tools==0.16.0
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
triton==2.3.1
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.49
ultralytics-thop==2.0.0
urllib3==2.2.2
webcolors==1.11.1
Werkzeug==3.0.3
yolov5==7.0.13
